Remove IdentityResolver from WebSocket use cases
